requirements.txt
streamlit>=1.32
google-genai>=1.0
